Fix function call register exhaustion issue
Previously, when we called a function, we evaluated function arguments
from left to right and kept the results in registers. That means the
number of available temporary registers decreased as we evaluate
arguments. It was very easy to see "register exhausted" error due to
that design.

This patch mitigates the issue. Now, we transform a function call
`f(a,b,...)` as follows:

t1=a
t2=b
...
f(t1, t2, ...)