Add a script for stage2 build
self.sh modifies source files for zcc so that they are
more friendly for self-hosting. The script is used to create
a zcc binary that consists of object files partly generated
by gcc and the others by zcc.

zcc can now compile main.c and type.c whith the script.