Emit position-independent code by default
So far, we have been using only static linking. In static linking, all
library functions (such as "printf") that an executable uses are
copied to the executable file. At runtime, library functions are just
a part of the executable file.

As opposed to static linking, there's a mechanism called "dynamic
linking". in dynamic linking, library functions are kept separated at
static-link time. At runtime, both the executable file and the library
files that the executable depends are loaded to memory, and the
dynamic linker links them in memory.

The static and dynamic linkers work hard to hide the details of the
dynamic linking, but that's not perfect. If you want to use dynamic
linking, your compiler has to emit dynamic linking-friendly assembly.
This patch does that.